web/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
pymongo==3.12.0
celery==5.1.2
redis==3.5.3
python-dotenv==0.19.0